move takes int

+++ src/flap.t.sol
@@ -42,8 +42,8 @@ contract WarpFlap is Flapper {
 
 contract VatLike is DSToken('') {
     uint constant ONE = 10 ** 27;
-    function move(bytes32 src, bytes32 dst, uint wad) public {
-        move(address(src), address(dst), wad / ONE);
+    function move(bytes32 src, bytes32 dst, int wad) public {
+        move(address(src), address(dst), uint(wad) / ONE);
     }
 }
